LPS-170680 Reorganize files into new directories
