+++ modules/apps/notification/notification-web/src/main/resources/META-INF/resources/js/components/BasicInfoContainer/BasicInfoContainer.tsx
