+++ modules/apps/notification/notification-web/src/main/resources/META-INF/resources/js/components/ContentContainer/Attachments.scss
