+++ modules/apps/notification/notification-web/src/main/resources/META-INF/resources/js/components/ContentContainer/DefinitionOfTerms.tsx
@@ -22,7 +22,7 @@ import {
 import {createResourceURL, fetch} from 'frontend-js-web';
 import React, {useEffect, useMemo, useState} from 'react';
 
-import {defaultLanguageId} from '../util/constants';
+import {defaultLanguageId} from '../../util/constants';
 
 interface DefinitionOfTermsProps {
 	baseResourceURL: string;
