+++ modules/apps/notification/notification-web/src/main/resources/META-INF/resources/js/components/SettingsContainer/EmailNotificationSettings.tsx
@@ -19,8 +19,7 @@ import {
 } from '@liferay/object-js-components-web';
 import React from 'react';
 
-import './EditNotificationTemplate.scss';
-import {NotificationTemplateError} from './EditNotificationTemplate';
+import {NotificationTemplateError} from '../EditNotificationTemplate';
 
 interface EmailNotificationSettingsProps {
 	errors: FormError<NotificationTemplate & NotificationTemplateError>;
