+++ modules/apps/notification/notification-web/src/main/resources/META-INF/resources/js/types.d.ts
@@ -38,6 +38,7 @@ interface NotificationTemplate {
 	body: LocalizedValue<string>;
 	description: string;
 	editorType: editorTypeOptions;
+	externalReferenceCode: string;
 	name: string;
 	objectDefinitionExternalReferenceCode: string;
 	objectDefinitionId: number | null;
