+++ modules/apps/notification/notification-web/types/src/main/resources/META-INF/resources/js/components/ContentContainer/FreeMarkerTemplateEditor.d.ts
