+++ modules/apps/notification/notification-web/types/src/main/resources/META-INF/resources/js/components/SettingsContainer/EmailNotificationSettings.d.ts
@@ -15,18 +15,17 @@
 /// <reference types="react" />
 
 import {FormError} from '@liferay/object-js-components-web';
-import {NotificationTemplateError} from './EditNotificationTemplate';
-import './EditNotificationTemplate.scss';
-interface SettingsContainerProps {
+import {NotificationTemplateError} from '../EditNotificationTemplate';
+interface EmailNotificationSettingsProps {
 	errors: FormError<NotificationTemplate & NotificationTemplateError>;
 	selectedLocale: Locale;
 	setValues: (values: Partial<NotificationTemplate>) => void;
 	values: NotificationTemplate;
 }
-export declare function SettingsContainer({
+export declare function EmailNotificationSettings({
 	errors,
 	selectedLocale,
 	setValues,
 	values,
-}: SettingsContainerProps): JSX.Element;
+}: EmailNotificationSettingsProps): JSX.Element;
 export {};
