+++ modules/apps/notification/notification-web/types/src/main/resources/META-INF/resources/js/components/SettingsContainer/UserNotificationSettings.d.ts
@@ -14,7 +14,6 @@
 
 /// <reference types="react" />
 
-import './EditNotificationTemplate.scss';
 interface UserNotificationSettingsProps {
 	setValues: (values: Partial<NotificationTemplate>) => void;
 	values: NotificationTemplate;
